% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_hashtags.R
\name{extract_hashtags}
\alias{extract_hashtags}
\title{Extracts all hashtags included in the input tweet data list}
\usage{
extract_hashtags(tweets)
}
\arguments{
\item{tweets}{A vector of tweets}
}
\value{
Vector of hashtags
}
\description{
Extracts all hashtags included in the input tweet data list
}
\examples{
tweets <- c(
   "Make America Great Again #DonaldTrump"
)
extract_hashtags(tweets)
>[1] "DonaldTrump"
}
