% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ngram.R
\name{extract_ngram}
\alias{extract_ngram}
\title{Extract n-grams from the input tweets data}
\usage{
extract_ngram(tweets, n = 3)
}
\arguments{
\item{tweets}{A vector of tweets}

\item{n}{Length of n-grams to be created
Default is 3. n-grams of length 3 will be created}
}
\value{
Vector of n-grams
}
\description{
Extract n-grams from the input tweets data
}
\examples{
tweets <- c(
   "Make", "America", "Great", "Again", "DonaldTrump"
)
extract_ngram(tweets, n=3)
}
