% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cloud.R
\name{generate_cloud}
\alias{generate_cloud}
\title{Generate word cloud}
\usage{
generate_cloud(tweets, type = "words")
}
\arguments{
\item{tweets}{A vector of tweets}

\item{type}{One of {"words", "hashtag", "stopwords"}, default="words".
Type of content to show in wordcloud. "words" shows all words,
"hashtag" only shows hashtags,
and "stopwords" does not show common English words.}
}
\value{
A plot figure
}
\description{
Generate word cloud
}
\examples{
tweets <- c("Make America Great Again! @DonaldTrump #America",
"It's rocket-science tier investment~~ #LoveElonMusk",
"America America America always GREAT",
"make america great again!")
cloud_df <- generate_cloud(tweets)
}
