% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_punct.R
\name{remove_punct}
\alias{remove_punct}
\title{Remove all punctuation and special characters from the
input tweets data}
\usage{
remove_punct(tweets, skip = NULL)
}
\arguments{
\item{tweets}{A vector of tweets}

\item{skip}{Vector of charcters or NULL.
The set of characters that do not have to be removed.
Default is NULL. If NULL, all characters except alphabets,
numbers and space would be removed.}
}
\value{
Vector of tweets without special characters
}
\description{
Remove all punctuation and special characters from the
input tweets data
}
\examples{
tweets <- c(
   "Make America Great Again! @DonaldTrump",
    "It's rocket-science tier investment~~ #LoveElonMusk"
)
remove_punct(tweets)
remove_punct(tweets, skip = c("'", "@", "#", "-"))
}
